+++ dipimage/writeics.cpp
@@ -40,7 +40,7 @@ void mexFunction( int /*nlhs*/, mxArray* /*plhs*/[], int nrhs, const mxArray* pr
       if( nrhs > 3 ) {
          significantBits = dml::GetUnsigned( prhs[ 3 ] );
       }
-      dip::StringSet options = {};
+      dip::StringSet options = { "fast" };
       if( nrhs > 4 ) {
          options = dml::GetStringSet( prhs[ 4 ] );
       }