+++ test/test_file_io.cpp
@@ -6,36 +6,78 @@
 #include <diplib/testing.h>
 
 int main() {
-   // Test ICS, 3D grey-value
 
-   std::cout << "\nTEST ICS\n\n";
+   // Test ICS, 3D grey-value
+   std::cout << "\nTEST ICS -- standard strides\n\n";
    dip::Image image = dip::ImageReadICS( "../test/chromo3d.ics" );
    image.SetPixelSize( dip::PhysicalQuantityArray{ 6 * dip::Units::Micrometer(), 300 * dip::Units::Nanometer() } );
-   std::cout << "Input image: " << image << std::endl;
 
    dip::testing::Timer timer;
    dip::ImageWriteICS( image, "test1.ics", { "line1", "line2 is good" }, 7, { "v1", "gzip" } );
    timer.Stop();
    std::cout << "Writing: " << timer << std::endl;
    timer.Reset();
-   dip::Image result = dip::ImageReadICS( "test1" );
+   dip::Image result = dip::ImageReadICS( "test1", dip::RangeArray{}, {} );
    timer.Stop();
    std::cout << "Reading: " << timer << std::endl;
-   std::cout << "Image read back: " << result;
+   std::cout << ( dip::testing::CompareImages( image, result ) ? "Identical\n\n" : "!!!ERROR!!!\n\n" );
+
+   timer.Reset();
+   dip::ImageWriteICS( image, "test1f.ics", { "line1", "line2 is good" }, 7, { "v1", "gzip", "fast" } );
+   timer.Stop();
+   std::cout << "Writing (fast): " << timer << std::endl;
+   timer.Reset();
+   result = dip::ImageReadICS( "test1f", dip::RangeArray{}, {}, "fast" );
+   timer.Stop();
+   std::cout << "Reading (fast): " << timer << std::endl;
+   std::cout << ( dip::testing::CompareImages( image, result ) ? "Identical\n\n" : "!!!ERROR!!!\n\n" );
+
+   timer.Reset();
+   result = dip::ImageReadICS( "test1f", dip::RangeArray{}, {} );
+   timer.Stop();
+   std::cout << "Reading (fast file): " << timer << std::endl;
+   std::cout << ( dip::testing::CompareImages( image, result ) ? "Identical\n\n" : "!!!ERROR!!!\n\n" );
+
+   timer.Reset();
+   result = dip::ImageReadICS( "test1", dip::RangeArray{}, {}, "fast" );
+   timer.Stop();
+   std::cout << "Reading (fast, regular file): " << timer << std::endl;
    std::cout << ( dip::testing::CompareImages( image, result ) ? "Identical\n\n" : "!!!ERROR!!!\n\n" );
 
    // Turn it on its side so the image to write has non-standard strides
-   image.SwapDimensions( 0, 1 );
-   std::cout << "Input image with non-standard strides: " << image;
+   std::cout << "\nTEST ICS -- non-standard strides\n\n";
+   image.SwapDimensions( 0, 2 );
+
    timer.Reset();
    dip::ImageWriteICS( image, "test2.ics", { "key\tvalue" }, 7, { "v1", "gzip" } );
    timer.Stop();
    std::cout << "Writing: " << timer << std::endl;
    timer.Reset();
-   dip::ImageReadICS( result, "test2" );
+   result = dip::ImageReadICS( "test2", dip::RangeArray{}, {} );
    timer.Stop();
    std::cout << "Reading: " << timer << std::endl;
-   std::cout << "Image read back: " << result;
+   std::cout << ( dip::testing::CompareImages( image, result ) ? "Identical\n\n" : "!!!ERROR!!!\n\n" );
+
+   timer.Reset();
+   dip::ImageWriteICS( image, "test2f.ics", { "key\tvalue" }, 7, { "v1", "gzip", "fast" } );
+   timer.Stop();
+   std::cout << "Writing (fast): " << timer << std::endl;
+   timer.Reset();
+   result = dip::ImageReadICS( "test2f", dip::RangeArray{}, {}, "fast" );
+   timer.Stop();
+   std::cout << "Reading (fast): " << timer << std::endl;
+   std::cout << ( dip::testing::CompareImages( image, result ) ? "Identical\n\n" : "!!!ERROR!!!\n\n" );
+
+   timer.Reset();
+   result = dip::ImageReadICS( "test2f", dip::RangeArray{}, {} );
+   timer.Stop();
+   std::cout << "Reading (fast file): " << timer << std::endl;
+   std::cout << ( dip::testing::CompareImages( image, result ) ? "Identical\n\n" : "!!!ERROR!!!\n\n" );
+
+   timer.Reset();
+   result = dip::ImageReadICS( "test2", dip::RangeArray{}, {}, "fast" );
+   timer.Stop();
+   std::cout << "Reading (fast, regular file): " << timer << std::endl;
    std::cout << ( dip::testing::CompareImages( image, result ) ? "Identical\n\n" : "!!!ERROR!!!\n\n" );
 
    // Test TIFF, 2D grey-value